make clang compile

+++ libft/libft_strstd/src/alloc/ft_strsetnew.c
@@ -3,19 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strsetnew.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jpriou <jpriou@student.42.fr>              +#+  +:+       +#+        */
+/*   By: pepe <pepe@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/12 18:10:52 by jpriou            #+#    #+#             */
-/*   Updated: 2017/11/12 18:13:41 by jpriou           ###   ########.fr       */
+/*   Updated: 2019/08/13 16:30:54 by pepe             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft_strstd.h"
 
-char		*ft_strsetnew(int len, char c)
+char		*ft_strsetnew(size_t len, char c)
 {
 	char	*res;
-	int		i;
+	size_t	i;
 
 	res = ft_strnew(len);
 	i = 0;
